+++ server/server.py
@@ -1,7 +1,7 @@
 from flask import Flask, request, jsonify, render_template, session
 import copy
 import os
-import dm_ai
+import DearMe_BE.server.ai as ai
 
 app = Flask(__name__)
 app.secret_key = os.urandom(24)
@@ -10,13 +10,13 @@ app.secret_key = os.urandom(24)
 def index():
     session.pop('chat', None)
     session['chat'] = session.get('chat', copy.deepcopy(conversation_history))
-    return render_template('newIndex.html')
+    return render_template('index.html')
 
 @app.route('/submit_form', methods=['POST'])
 def submit_form():
     global client
 
-    response_message = dm_ai.generate_chat(client, request.form['message'], session['chat'])
+    response_message = ai.generate_chat(client, request.form['message'], session['chat'])
     session.modified = True
     print("Chat history:", session['chat'])
 
@@ -27,8 +27,8 @@ def submit_form():
     return jsonify(response_data)
 
 if __name__ == '__main__':
-    client = dm_ai.create_openai_client()
-    conversation_history = [{"role": "system", "content": dm_ai.system_prompt}]
-    dm_ai.system_token = dm_ai.num_tokens_from_messages(conversation_history, model=dm_ai.MODEL)
-    dm_ai.encoding = dm_ai.tiktoken.encoding_for_model(dm_ai.MODEL)
+    client = ai.create_openai_client()
+    conversation_history = [{"role": "system", "content": ai.system_prompt}]
+    ai.system_token = ai.num_tokens_from_messages(conversation_history, model=ai.MODEL)
+    ai.encoding = ai.tiktoken.encoding_for_model(ai.MODEL)
     app.run(debug=True)
